server/stable/json/latest.json ==============================
This file represents the available variables to describe a server aka core aka OC10 release.
The file models the structure visible in wordpress as json.

The section "main" is the table next to the text block "ownCloud 10 Source packages"
The section "minimal" is normally collapsed behind "Need a minimal version to ...?"
Sections "qa" and "qa-minimal" are not shown. But they exist in the download area, so I've added that for completeness.
Sections "docker", "linux", and "appliance" all have just one or emultiple buttons. I've providedd lists
of label + url for these buttons, in the order of appearance.

This latest.json only represents one release.
Historic versions could be additional files with version number or timestamp a name.
We need to find out, if that is needed here.